Add write access check for contests endpoints

+++ src/main/java/com/ensiasit/projectx/services/ContestServiceImpl.java
@@ -2,14 +2,17 @@ package com.ensiasit.projectx.services;
 
 import com.ensiasit.projectx.dto.ContestDto;
 import com.ensiasit.projectx.dto.UserContestRoleDto;
+import com.ensiasit.projectx.exceptions.ForbiddenException;
 import com.ensiasit.projectx.exceptions.NotFoundException;
 import com.ensiasit.projectx.mappers.ContestMapper;
 import com.ensiasit.projectx.models.Contest;
+import com.ensiasit.projectx.models.UserContestRole;
 import com.ensiasit.projectx.repositories.ContestRepository;
 import com.ensiasit.projectx.repositories.UserContestRoleRepository;
 import com.ensiasit.projectx.utils.RoleEnum;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -30,7 +33,11 @@ public class ContestServiceImpl implements ContestService {
     }
 
     @Override
-    public ContestDto createContest(ContestDto contestDto) {
+    public ContestDto createContest(String userEmail, ContestDto contestDto) {
+        if (!adminService.isAdmin(userEmail)) {
+            throw new ForbiddenException("User has no write access.");
+        }
+
         Contest contest = contestRepository.save(contestMapper.fromContestDto(contestDto));
 
         return contestMapper.toContestDto(contest);
@@ -43,9 +50,16 @@ public class ContestServiceImpl implements ContestService {
         return contestMapper.toContestDto(contest);
     }
 
+    @Transactional
     @Override
-    public ContestDto deleteContest(long id) {
+    public ContestDto deleteContest(String userEmail, long id) {
         Contest contest = extractContest(id);
+
+        if (userHasNoWriteAccess(userEmail, id)) {
+            throw new ForbiddenException("User has no write access.");
+        }
+
+        userContestRoleRepository.deleteAllByContestId(id);
         contestRepository.deleteById(id);
 
         return contestMapper.toContestDto(contest);
@@ -65,8 +79,12 @@ public class ContestServiceImpl implements ContestService {
     }
 
     @Override
-    public ContestDto updateContest(long id, ContestDto payload) {
+    public ContestDto updateContest(String userEmail, long id, ContestDto payload) {
         Contest contest = extractContest(id);
+
+        if (userHasNoWriteAccess(userEmail, id)) {
+            throw new ForbiddenException("User has no write access.");
+        }
 
         contest.setName(payload.getName());
         contest.setStartTime(payload.getStartTime());
@@ -95,5 +113,18 @@ public class ContestServiceImpl implements ContestService {
         }
 
         return contestOptional.get();
+    }
+
+    private boolean userHasNoWriteAccess(String userEmail, long contestId) {
+        if (adminService.isAdmin(userEmail)) {
+            return false;
+        }
+
+        List<UserContestRole> userContestRoles = userContestRoleRepository.findAllByUserEmail(userEmail);
+
+        return userContestRoles.stream()
+                .filter(userContestRole -> userContestRole.getContest().getId() == contestId)
+                .noneMatch(userContestRole ->
+                        userContestRole.getRole().equals(RoleEnum.ROLE_ADMIN) || userContestRole.getRole().equals(RoleEnum.ROLE_MODERATOR));
     }
 }
